app/requirements.txt
# LibraryExplorer Dependencies

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6  # For file uploads

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Math & Scientific Computing
numpy>=1.26.0
scipy>=1.12.0

# Dimensionality Reduction & Clustering
umap-learn>=0.5.5
scikit-learn>=1.4.0

# Graph Operations
networkx>=3.2.0
python-louvain>=0.16    # community detection

# Configuration
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0           # For FastAPI TestClient

# Development (optional)
# black>=24.0.0
# ruff>=0.1.0
# mypy>=1.8.0
